Undo file reorg and renaming
