+++ azure-cli.pyproj
@@ -1150,7 +1150,7 @@
     <Content Include="command_modules\azure-cli-batch\azure\cli\command_modules\batch\tests\latest\README.md" />
     <Content Include="command_modules\azure-cli-batch\HISTORY.rst" />
     <Content Include="command_modules\azure-cli-botservice\azure\.vscode\tags" />
-    <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\bot\templates\functionapp.template.json" />
+    <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\functionapp.template.json" />
     <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\tests\latest\recordings\test_auth_setting.yaml" />
     <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\tests\latest\recordings\test_directline_channel.yaml" />
     <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\tests\latest\recordings\test_msteams_channel.yaml" />
@@ -1159,16 +1159,16 @@
     <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\tests\latest\recordings\test_skype_channel.yaml" />
     <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\tests\latest\recordings\test_webapp_bot.yaml" />
     <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\tests\latest\recordings\test_webchat_channel.yaml" />
-    <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\bot\templates\webapp.template.json" />
-    <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\bot\templates\webappv4.template.json" />
+    <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\webapp.template.json" />
+    <Content Include="command_modules\azure-cli-botservice\azure\cli\command_modules\botservice\webappv4.template.json" />
     <Content Include="command_modules\azure-cli-botservice\azure_cli_botservice.egg-info\dependency_links.txt" />
     <Content Include="command_modules\azure-cli-botservice\azure_cli_botservice.egg-info\PKG-INFO" />
     <Content Include="command_modules\azure-cli-botservice\azure_cli_botservice.egg-info\requires.txt" />
     <Content Include="command_modules\azure-cli-botservice\azure_cli_botservice.egg-info\SOURCES.txt" />
     <Content Include="command_modules\azure-cli-botservice\azure_cli_botservice.egg-info\top_level.txt" />
-    <Content Include="command_modules\azure-cli-botservice\build\lib.linux-x86_64-2.7\azure\cli\command_modules\botservice\bot\templates\functionapp.template.json" />
-    <Content Include="command_modules\azure-cli-botservice\build\lib.linux-x86_64-2.7\azure\cli\command_modules\botservice\bot\templates\webapp.template.json" />
-    <Content Include="command_modules\azure-cli-botservice\build\lib.linux-x86_64-2.7\azure\cli\command_modules\botservice\bot\templates\webappv4.template.json" />
+    <Content Include="command_modules\azure-cli-botservice\build\lib.linux-x86_64-2.7\azure\cli\command_modules\botservice\functionapp.template.json" />
+    <Content Include="command_modules\azure-cli-botservice\build\lib.linux-x86_64-2.7\azure\cli\command_modules\botservice\webapp.template.json" />
+    <Content Include="command_modules\azure-cli-botservice\build\lib.linux-x86_64-2.7\azure\cli\command_modules\botservice\webappv4.template.json" />
     <Content Include="command_modules\azure-cli-botservice\dist\azure_cli_botservice-0.1.0-py2.py3-none-any.whl" />
     <Content Include="command_modules\azure-cli-botservice\HISTORY.rst" />
     <Content Include="command_modules\azure-cli-botservice\MANIFEST.in" />
