+++ src/command_modules/azure-cli-botservice/MANIFEST.in
@@ -1,4 +1,4 @@
 include *.rst
-include azure/cli/command_modules/botservice/bot/templates/functionapp.template.json
-include azure/cli/command_modules/botservice/bot/templates/webapp.template.json
-include azure/cli/command_modules/botservice/bot/templates/webappv4.template.json
+include azure/cli/command_modules/botservice/functionapp.template.json
+include azure/cli/command_modules/botservice/webapp.template.json
+include azure/cli/command_modules/botservice/webappv4.template.json
