+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/adal_authenticator.py
