+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/azure_region_mapper.py
