+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/bot_publish_prep.py
